+++ components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Protocol/src/Services/ClientUtils.cs
@@ -1,0 +1,666 @@
+﻿namespace Microsoft.Azure.IIoT.OpcUa.Protocol.Services {
+    using Opc.Ua;
+    using Opc.Ua.Client;
+    using System;
+    using System.Collections.Generic;
+
+    internal static class ClientUtils {
+        #region Browse
+        /// <summary>
+        /// Browses the address space and returns the references found.
+        /// </summary>
+        /// <param name="session">The session.</param>
+        /// <param name="nodesToBrowse">The set of browse operations to perform.</param>
+        /// <param name="throwOnError">if set to <c>true</c> a exception will be thrown on an error.</param>
+        /// <returns>
+        /// The references found. Null if an error occurred.
+        /// </returns>
+        public static ReferenceDescriptionCollection Browse(Session session, BrowseDescriptionCollection nodesToBrowse, bool throwOnError) {
+            return Browse(session, null, nodesToBrowse, throwOnError);
+        }
+
+        /// <summary>
+        /// Browses the address space and returns the references found.
+        /// </summary>
+        public static ReferenceDescriptionCollection Browse(Session session, ViewDescription view, BrowseDescriptionCollection nodesToBrowse, bool throwOnError) {
+            try {
+                ReferenceDescriptionCollection references = new ReferenceDescriptionCollection();
+                BrowseDescriptionCollection unprocessedOperations = new BrowseDescriptionCollection();
+
+                while (nodesToBrowse.Count > 0) {
+                    // start the browse operation.
+                    BrowseResultCollection results = null;
+                    DiagnosticInfoCollection diagnosticInfos = null;
+
+                    session.Browse(
+                        null,
+                        view,
+                        0,
+                        nodesToBrowse,
+                        out results,
+                        out diagnosticInfos);
+
+                    ClientBase.ValidateResponse(results, nodesToBrowse);
+                    ClientBase.ValidateDiagnosticInfos(diagnosticInfos, nodesToBrowse);
+
+                    ByteStringCollection continuationPoints = new ByteStringCollection();
+
+                    for (int ii = 0; ii < nodesToBrowse.Count; ii++) {
+                        // check for error.
+                        if (StatusCode.IsBad(results[ii].StatusCode)) {
+                            // this error indicates that the server does not have enough simultaneously active 
+                            // continuation points. This request will need to be resent after the other operations
+                            // have been completed and their continuation points released.
+                            if (results[ii].StatusCode == StatusCodes.BadNoContinuationPoints) {
+                                unprocessedOperations.Add(nodesToBrowse[ii]);
+                            }
+
+                            continue;
+                        }
+
+                        // check if all references have been fetched.
+                        if (results[ii].References.Count == 0) {
+                            continue;
+                        }
+
+                        // save results.
+                        references.AddRange(results[ii].References);
+
+                        // check for continuation point.
+                        if (results[ii].ContinuationPoint != null) {
+                            continuationPoints.Add(results[ii].ContinuationPoint);
+                        }
+                    }
+
+                    // process continuation points.
+                    ByteStringCollection revisedContiuationPoints = new ByteStringCollection();
+
+                    while (continuationPoints.Count > 0) {
+                        // continue browse operation.
+                        session.BrowseNext(
+                            null,
+                            false,
+                            continuationPoints,
+                            out results,
+                            out diagnosticInfos);
+
+                        ClientBase.ValidateResponse(results, continuationPoints);
+                        ClientBase.ValidateDiagnosticInfos(diagnosticInfos, continuationPoints);
+
+                        for (int ii = 0; ii < continuationPoints.Count; ii++) {
+                            // check for error.
+                            if (StatusCode.IsBad(results[ii].StatusCode)) {
+                                continue;
+                            }
+
+                            // check if all references have been fetched.
+                            if (results[ii].References.Count == 0) {
+                                continue;
+                            }
+
+                            // save results.
+                            references.AddRange(results[ii].References);
+
+                            // check for continuation point.
+                            if (results[ii].ContinuationPoint != null) {
+                                revisedContiuationPoints.Add(results[ii].ContinuationPoint);
+                            }
+                        }
+
+                        // check if browsing must continue;
+                        revisedContiuationPoints = continuationPoints;
+                    }
+
+                    // check if unprocessed results exist.
+                    nodesToBrowse = unprocessedOperations;
+                }
+
+                // return complete list.
+                return references;
+            }
+            catch (Exception exception) {
+                if (throwOnError) {
+                    throw new ServiceResultException(exception, StatusCodes.BadUnexpectedError);
+                }
+
+                return null;
+            }
+        }
+
+        /// <summary>
+        /// Browses the address space and returns the references found.
+        /// </summary>
+        /// <param name="session">The session.</param>
+        /// <param name="nodeToBrowse">The NodeId for the starting node.</param>
+        /// <param name="throwOnError">if set to <c>true</c> a exception will be thrown on an error.</param>
+        /// <returns>
+        /// The references found. Null if an error occurred.
+        /// </returns>
+        public static ReferenceDescriptionCollection Browse(Session session, BrowseDescription nodeToBrowse, bool throwOnError) {
+            return Browse(session, null, nodeToBrowse, throwOnError);
+        }
+
+        /// <summary>
+        /// Browses the address space and returns the references found.
+        /// </summary>
+        public static ReferenceDescriptionCollection Browse(Session session, ViewDescription view, BrowseDescription nodeToBrowse, bool throwOnError) {
+            try {
+                ReferenceDescriptionCollection references = new ReferenceDescriptionCollection();
+
+                // construct browse request.
+                BrowseDescriptionCollection nodesToBrowse = new BrowseDescriptionCollection();
+                nodesToBrowse.Add(nodeToBrowse);
+
+                // start the browse operation.
+                BrowseResultCollection results = null;
+                DiagnosticInfoCollection diagnosticInfos = null;
+
+                session.Browse(
+                    null,
+                    view,
+                    0,
+                    nodesToBrowse,
+                    out results,
+                    out diagnosticInfos);
+
+                ClientBase.ValidateResponse(results, nodesToBrowse);
+                ClientBase.ValidateDiagnosticInfos(diagnosticInfos, nodesToBrowse);
+
+                do {
+                    // check for error.
+                    if (StatusCode.IsBad(results[0].StatusCode)) {
+                        throw new ServiceResultException(results[0].StatusCode);
+                    }
+
+                    // process results.
+                    for (int ii = 0; ii < results[0].References.Count; ii++) {
+                        references.Add(results[0].References[ii]);
+                    }
+
+                    // check if all references have been fetched.
+                    if (results[0].References.Count == 0 || results[0].ContinuationPoint == null) {
+                        break;
+                    }
+
+                    // continue browse operation.
+                    ByteStringCollection continuationPoints = new ByteStringCollection();
+                    continuationPoints.Add(results[0].ContinuationPoint);
+
+                    session.BrowseNext(
+                        null,
+                        false,
+                        continuationPoints,
+                        out results,
+                        out diagnosticInfos);
+
+                    ClientBase.ValidateResponse(results, continuationPoints);
+                    ClientBase.ValidateDiagnosticInfos(diagnosticInfos, continuationPoints);
+                }
+                while (true);
+
+                //return complete list.
+                return references;
+            }
+            catch (Exception exception) {
+                if (throwOnError) {
+                    throw new ServiceResultException(exception, StatusCodes.BadUnexpectedError);
+                }
+
+                return null;
+            }
+        }
+
+        /// <summary>
+        /// Browses the address space and returns all of the supertypes of the specified type node.
+        /// </summary>
+        /// <param name="session">The session.</param>
+        /// <param name="typeId">The NodeId for a type node in the address space.</param>
+        /// <param name="throwOnError">if set to <c>true</c> a exception will be thrown on an error.</param>
+        /// <returns>
+        /// The references found. Null if an error occurred.
+        /// </returns>
+        public static ReferenceDescriptionCollection BrowseSuperTypes(Session session, NodeId typeId, bool throwOnError) {
+            ReferenceDescriptionCollection supertypes = new ReferenceDescriptionCollection();
+
+            try {
+                // find all of the children of the field.
+                BrowseDescription nodeToBrowse = new BrowseDescription();
+
+                nodeToBrowse.NodeId = typeId;
+                nodeToBrowse.BrowseDirection = Opc.Ua.BrowseDirection.Inverse;
+                nodeToBrowse.ReferenceTypeId = ReferenceTypeIds.HasSubtype;
+                nodeToBrowse.IncludeSubtypes = false; // more efficient to use IncludeSubtypes=False when possible.
+                nodeToBrowse.NodeClassMask = 0; // the HasSubtype reference already restricts the targets to Types. 
+                nodeToBrowse.ResultMask = (uint)BrowseResultMask.All;
+
+                ReferenceDescriptionCollection references = Browse(session, nodeToBrowse, throwOnError);
+
+                while (references != null && references.Count > 0) {
+                    // should never be more than one supertype.
+                    supertypes.Add(references[0]);
+
+                    // only follow references within this server.
+                    if (references[0].NodeId.IsAbsolute) {
+                        break;
+                    }
+
+                    // get the references for the next level up.
+                    nodeToBrowse.NodeId = (NodeId)references[0].NodeId;
+                    references = Browse(session, nodeToBrowse, throwOnError);
+                }
+
+                // return complete list.
+                return supertypes;
+            }
+            catch (Exception exception) {
+                if (throwOnError) {
+                    throw new ServiceResultException(exception, StatusCodes.BadUnexpectedError);
+                }
+
+                return null;
+            }
+        }
+
+        /// <summary>
+        /// Returns the node ids for a set of relative paths.
+        /// </summary>
+        /// <param name="session">An open session with the server to use.</param>
+        /// <param name="startNodeId">The starting node for the relative paths.</param>
+        /// <param name="namespacesUris">The namespace URIs referenced by the relative paths.</param>
+        /// <param name="relativePaths">The relative paths.</param>
+        /// <returns>A collection of local nodes.</returns>
+        public static List<NodeId> TranslateBrowsePaths(
+            Session session,
+            NodeId startNodeId,
+            NamespaceTable namespacesUris,
+            params string[] relativePaths) {
+            // build the list of browse paths to follow by parsing the relative paths.
+            BrowsePathCollection browsePaths = new BrowsePathCollection();
+
+            if (relativePaths != null) {
+                for (int ii = 0; ii < relativePaths.Length; ii++) {
+                    BrowsePath browsePath = new BrowsePath();
+
+                    // The relative paths used indexes in the namespacesUris table. These must be 
+                    // converted to indexes used by the server. An error occurs if the relative path
+                    // refers to a namespaceUri that the server does not recognize.
+
+                    // The relative paths may refer to ReferenceType by their BrowseName. The TypeTree object
+                    // allows the parser to look up the server's NodeId for the ReferenceType.
+
+                    browsePath.RelativePath = RelativePath.Parse(
+                        relativePaths[ii],
+                        session.TypeTree,
+                        namespacesUris,
+                        session.NamespaceUris);
+
+                    browsePath.StartingNode = startNodeId;
+
+                    browsePaths.Add(browsePath);
+                }
+            }
+
+            // make the call to the server.
+            BrowsePathResultCollection results;
+            DiagnosticInfoCollection diagnosticInfos;
+
+            ResponseHeader responseHeader = session.TranslateBrowsePathsToNodeIds(
+                null,
+                browsePaths,
+                out results,
+                out diagnosticInfos);
+
+            // ensure that the server returned valid results.
+            Session.ValidateResponse(results, browsePaths);
+            Session.ValidateDiagnosticInfos(diagnosticInfos, browsePaths);
+
+            // collect the list of node ids found.
+            List<NodeId> nodes = new List<NodeId>();
+
+            for (int ii = 0; ii < results.Count; ii++) {
+                // check if the start node actually exists.
+                if (StatusCode.IsBad(results[ii].StatusCode)) {
+                    nodes.Add(null);
+                    continue;
+                }
+
+                // an empty list is returned if no node was found.
+                if (results[ii].Targets.Count == 0) {
+                    nodes.Add(null);
+                    continue;
+                }
+
+                // Multiple matches are possible, however, the node that matches the type model is the
+                // one we are interested in here. The rest can be ignored.
+                BrowsePathTarget target = results[ii].Targets[0];
+
+                if (target.RemainingPathIndex != UInt32.MaxValue) {
+                    nodes.Add(null);
+                    continue;
+                }
+
+                // The targetId is an ExpandedNodeId because it could be node in another server. 
+                // The ToNodeId function is used to convert a local NodeId stored in a ExpandedNodeId to a NodeId.
+                nodes.Add(ExpandedNodeId.ToNodeId(target.TargetId, session.NamespaceUris));
+            }
+
+            // return whatever was found.
+            return nodes;
+        }
+        #endregion
+
+        #region Type Model Browsing
+
+        /// <summary>
+        /// Stores an instance declaration fetched from the server.
+        /// </summary>
+        public class InstanceDeclaration {
+            /// <summary>
+            /// The browse path to the instance declaration.
+            /// </summary>
+            public QualifiedNameCollection BrowsePath;
+
+            /// <summary>
+            /// The browse path to the instance declaration.
+            /// </summary>
+            public string BrowsePathDisplayText;
+
+            /// <summary>
+            /// The node id for the instance declaration.
+            /// </summary>
+            public NodeId NodeId;
+
+            /// <summary>
+            /// The node class of the instance declaration.
+            /// </summary>
+            public Opc.Ua.NodeClass NodeClass;
+
+            /// <summary>
+            /// The modelling rule for the instance declaration (i.e. Mandatory or Optional).
+            /// </summary>
+            public NodeId ModellingRule;
+
+            /// <summary>
+            /// The data type for the instance declaration.
+            /// </summary>
+            public NodeId DataType;
+
+            /// <summary>
+            /// The value rank for the instance declaration.
+            /// </summary>
+            public int ValueRank;
+        }
+
+
+        /// <summary>
+        /// Collects the instance declarations for a type.
+        /// </summary>
+        public static List<InstanceDeclaration> CollectInstanceDeclarationsForType(Session session, NodeId typeId) {
+            return CollectInstanceDeclarationsForType(session, typeId, true);
+        }
+
+        /// <summary>
+        /// Collects the instance declarations for a type.
+        /// </summary>
+        public static List<InstanceDeclaration> CollectInstanceDeclarationsForType(Session session, NodeId typeId, bool includeSupertypes) {
+            // process the types starting from the top of the tree.
+            List<InstanceDeclaration> instances = new List<InstanceDeclaration>();
+            Dictionary<string, InstanceDeclaration> map = new Dictionary<string, InstanceDeclaration>();
+
+            // get the supertypes.
+            if (includeSupertypes) {
+                ReferenceDescriptionCollection supertypes = BrowseSuperTypes(session, typeId, false);
+
+                if (supertypes != null) {
+                    for (int ii = supertypes.Count - 1; ii >= 0; ii--) {
+                        CollectInstanceDeclarations(session, (NodeId)supertypes[ii].NodeId, null, instances, map);
+                    }
+                }
+            }
+
+            // collect the fields for the selected type.
+            CollectInstanceDeclarations(session, typeId, null, instances, map);
+
+            // return the complete list.
+            return instances;
+        }
+
+        /// <summary>
+        /// Collects the fields for the instance node.
+        /// </summary>
+        private static void CollectInstanceDeclarations(
+            Session session,
+            NodeId typeId,
+            InstanceDeclaration parent,
+            List<InstanceDeclaration> instances,
+            IDictionary<string, InstanceDeclaration> map) {
+            // find the children.
+            BrowseDescription nodeToBrowse = new BrowseDescription();
+
+            if (parent == null) {
+                nodeToBrowse.NodeId = typeId;
+            }
+            else {
+                nodeToBrowse.NodeId = parent.NodeId;
+            }
+
+            nodeToBrowse.BrowseDirection = Opc.Ua.BrowseDirection.Forward;
+            nodeToBrowse.ReferenceTypeId = ReferenceTypeIds.HasChild;
+            nodeToBrowse.IncludeSubtypes = true;
+            nodeToBrowse.NodeClassMask = (uint)Opc.Ua.NodeClass.Variable;
+            nodeToBrowse.ResultMask = (uint)BrowseResultMask.All;
+
+            // ignore any browsing errors.
+            ReferenceDescriptionCollection references = Browse(session, nodeToBrowse, false);
+
+            if (references == null) {
+                return;
+            }
+
+            // process the children.
+            List<NodeId> nodeIds = new List<NodeId>();
+            List<InstanceDeclaration> children = new List<InstanceDeclaration>();
+
+            for (int ii = 0; ii < references.Count; ii++) {
+                ReferenceDescription reference = references[ii];
+
+                if (reference.NodeId.IsAbsolute) {
+                    continue;
+                }
+
+                // create a new declaration.
+                InstanceDeclaration child = new InstanceDeclaration();
+
+                child.NodeId = (NodeId)reference.NodeId;
+                child.NodeClass = reference.NodeClass;
+
+                if (parent != null) {
+                    child.BrowsePath = new QualifiedNameCollection(parent.BrowsePath);
+                    child.BrowsePathDisplayText = Utils.Format("{0}/{1}", parent.BrowsePathDisplayText, reference.BrowseName);
+                }
+                else {
+                    child.BrowsePath = new QualifiedNameCollection();
+                    child.BrowsePathDisplayText = Utils.Format("{0}", reference.BrowseName);
+                }
+
+                child.BrowsePath.Add(reference.BrowseName);
+
+                map[child.BrowsePathDisplayText] = child;
+
+                // add to list.
+                children.Add(child);
+                nodeIds.Add(child.NodeId);
+            }
+
+            // check if nothing more to do.
+            if (children.Count == 0) {
+                return;
+            }
+
+            // find the modelling rules.
+            List<NodeId> modellingRules = FindTargetOfReference(session, nodeIds, Opc.Ua.ReferenceTypeIds.HasModellingRule, false);
+
+            if (modellingRules != null) {
+                for (int ii = 0; ii < nodeIds.Count; ii++) {
+                    children[ii].ModellingRule = modellingRules[ii];
+
+                    // if the modelling rule is null then the instance is not part of the type declaration.
+                    if (NodeId.IsNull(modellingRules[ii])) {
+                        map.Remove(children[ii].BrowsePathDisplayText);
+                    }
+                }
+            }
+
+            // update the descriptions.
+            UpdateInstanceDescriptions(session, children, false);
+
+            // recusively collect instance declarations for the tree below.
+            for (int ii = 0; ii < children.Count; ii++) {
+                if (!NodeId.IsNull(children[ii].ModellingRule)) {
+                    instances.Add(children[ii]);
+                    CollectInstanceDeclarations(session, typeId, children[ii], instances, map);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Finds the targets for the specified reference.
+        /// </summary>
+        private static List<NodeId> FindTargetOfReference(Session session, List<NodeId> nodeIds, NodeId referenceTypeId, bool throwOnError) {
+            try {
+                // construct browse request.
+                BrowseDescriptionCollection nodesToBrowse = new BrowseDescriptionCollection();
+
+                for (int ii = 0; ii < nodeIds.Count; ii++) {
+                    BrowseDescription nodeToBrowse = new BrowseDescription();
+                    nodeToBrowse.NodeId = nodeIds[ii];
+                    nodeToBrowse.BrowseDirection = Opc.Ua.BrowseDirection.Forward;
+                    nodeToBrowse.ReferenceTypeId = referenceTypeId;
+                    nodeToBrowse.IncludeSubtypes = false;
+                    nodeToBrowse.NodeClassMask = 0;
+                    nodeToBrowse.ResultMask = (uint)BrowseResultMask.None;
+                    nodesToBrowse.Add(nodeToBrowse);
+                }
+
+                // start the browse operation.
+                BrowseResultCollection results = null;
+                DiagnosticInfoCollection diagnosticInfos = null;
+
+                session.Browse(
+                    null,
+                    null,
+                    1,
+                    nodesToBrowse,
+                    out results,
+                    out diagnosticInfos);
+
+                ClientBase.ValidateResponse(results, nodesToBrowse);
+                ClientBase.ValidateDiagnosticInfos(diagnosticInfos, nodesToBrowse);
+
+                List<NodeId> targetIds = new List<NodeId>();
+                ByteStringCollection continuationPoints = new ByteStringCollection();
+
+                for (int ii = 0; ii < nodeIds.Count; ii++) {
+                    targetIds.Add(null);
+
+                    // check for error.
+                    if (StatusCode.IsBad(results[ii].StatusCode)) {
+                        continue;
+                    }
+
+                    // check for continuation point.
+                    if (results[ii].ContinuationPoint != null && results[ii].ContinuationPoint.Length > 0) {
+                        continuationPoints.Add(results[ii].ContinuationPoint);
+                    }
+
+                    // get the node id.
+                    if (results[ii].References.Count > 0) {
+                        if (NodeId.IsNull(results[ii].References[0].NodeId) || results[ii].References[0].NodeId.IsAbsolute) {
+                            continue;
+                        }
+
+                        targetIds[ii] = (NodeId)results[ii].References[0].NodeId;
+                    }
+                }
+
+                // release continuation points.
+                if (continuationPoints.Count > 0) {
+                    session.BrowseNext(
+                        null,
+                        true,
+                        continuationPoints,
+                        out results,
+                        out diagnosticInfos);
+
+                    ClientBase.ValidateResponse(results, nodesToBrowse);
+                    ClientBase.ValidateDiagnosticInfos(diagnosticInfos, nodesToBrowse);
+                }
+
+                //return complete list.
+                return targetIds;
+            }
+            catch (Exception exception) {
+                if (throwOnError) {
+                    throw new ServiceResultException(exception, StatusCodes.BadUnexpectedError);
+                }
+
+                return null;
+            }
+        }
+
+        /// <summary>
+        /// Finds the targets for the specified reference.
+        /// </summary>
+        private static void UpdateInstanceDescriptions(Session session, List<InstanceDeclaration> instances, bool throwOnError) {
+            try {
+                ReadValueIdCollection nodesToRead = new ReadValueIdCollection();
+
+                for (int ii = 0; ii < instances.Count; ii++) {
+                    ReadValueId nodeToRead = new ReadValueId();
+                    nodeToRead.NodeId = instances[ii].NodeId;
+                    nodeToRead.AttributeId = Attributes.Description;
+                    nodesToRead.Add(nodeToRead);
+
+                    nodeToRead = new ReadValueId();
+                    nodeToRead.NodeId = instances[ii].NodeId;
+                    nodeToRead.AttributeId = Attributes.DataType;
+                    nodesToRead.Add(nodeToRead);
+
+                    nodeToRead = new ReadValueId();
+                    nodeToRead.NodeId = instances[ii].NodeId;
+                    nodeToRead.AttributeId = Attributes.ValueRank;
+                    nodesToRead.Add(nodeToRead);
+                }
+
+                // start the browse operation.
+                DataValueCollection results = null;
+                DiagnosticInfoCollection diagnosticInfos = null;
+
+                session.Read(
+                    null,
+                    0,
+                    TimestampsToReturn.Neither,
+                    nodesToRead,
+                    out results,
+                    out diagnosticInfos);
+
+                ClientBase.ValidateResponse(results, nodesToRead);
+                ClientBase.ValidateDiagnosticInfos(diagnosticInfos, nodesToRead);
+
+                // update the instances.
+                for (int ii = 0; ii < nodesToRead.Count; ii += 3) {
+                    InstanceDeclaration instance = instances[ii / 3];
+
+                    instance.DataType = results[ii + 1].GetValue<NodeId>(NodeId.Null);
+                    instance.ValueRank = results[ii + 2].GetValue<int>(ValueRanks.Any);
+                }
+            }
+            catch (Exception exception) {
+                if (throwOnError) {
+                    throw new ServiceResultException(exception, StatusCodes.BadUnexpectedError);
+                }
+            }
+        }
+        #endregion
+    }
+}
